Fix two more Windows-only RaceServer build errors
- stdafx.h never defined SocketLength on the _WIN32 branch (only the
  POSIX one had it), breaking accept()'s addrlen parameter.
- std::max(int, SOCKET) is ambiguous on Windows, where SOCKET is
  UINT_PTR rather than the plain int it's typedef'd to on Linux; cast
  to int at the one call site that compares a socket against a loop
  bound (ServerSocket.cpp:180).

Both are Windows-only branches/casts, so the Linux build and tests
are unaffected.

+++ NetTarget/RaceServer/NetworkInterface/ServerSocket.cpp
@@ -177,7 +177,7 @@ void MR_ServerSocket::ProcessEvents(MR_RaceManager* pRaceManager)
     int maxSocket = mDatagramSocket;
     for (const auto& pair : mConnections) {
         if (pair.second && pair.second->mConnected) {
-            maxSocket = std::max(maxSocket, pair.second->mTcpSocket);
+            maxSocket = std::max(maxSocket, static_cast<int>(pair.second->mTcpSocket));
         }
     }
     int selectResult = select(maxSocket + 1, &readSet, NULL, NULL, &timeout);
